requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-docx==0.8.11
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3  # Required for better BeautifulSoup performance
python-magic==0.4.27  # For file type detection

# Pro Tier Dependencies
spacy==3.7.1
scikit-learn==1.3.2
openai==1.3.7
numpy==1.26.2
